refactor: move CardSetting to submodule #37
