+++ src/Components/Common/Card/Card.style.tsx
@@ -101,10 +101,3 @@ export const SubmitForm = styled.form`
   gap: 0.5rem;
   height: 2.75rem;
 `;
-
-export const MenuWrapper = styled.div`
-  position: absolute;
-  top: 0.5rem;
-  right: 0.5rem;
-  z-index: 1;
-`;
