+++ src/Components/Common/Card/CardSetting/CardSetting.style.tsx
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+export const MenuWrapper = styled.div`
+  position: absolute;
+  top: 0.5rem;
+  right: 0.5rem;
+  z-index: 1;
+`;
